# Switch to single-transaction native SQL compaction