--- a/pkg/drivers/mysql/mysql.go
+++ b/pkg/drivers/mysql/mysql.go
@@ -65,7 +65,26 @@ func New(ctx context.Context, dataSourceName string, tlsInfo tls.Config, connPoo
 	if err != nil {
 		return nil, err
 	}
+
 	dialect.LastInsertID = true
+	dialect.CompactSQL = `
+		DELETE kv FROM kine AS kv
+		INNER JOIN (
+			SELECT kp.prev_revision AS id
+			FROM kine AS kp
+			WHERE
+				kp.prev_revision != 0 AND
+				kp.id <= ?
+			UNION
+			SELECT kd.id AS id
+			FROM kine AS kd
+			WHERE
+				kd.deleted != 0 AND
+				kd.id <= ?
+		) AS ks
+		ON
+			kv.id = ks.id AND
+			kv.name != 'compact_rev_key'`
 	dialect.TranslateErr = func(err error) error {
 		if err, ok := err.(*mysql.MySQLError); ok && err.Number == 1062 {
 			return server.ErrKeyExists