--- a/pkg/drivers/pgsql/pgsql.go
+++ b/pkg/drivers/pgsql/pgsql.go
@@ -57,6 +57,23 @@ func New(ctx context.Context, dataSourceName string, tlsInfo tls.Config, connPoo
 	if err != nil {
 		return nil, err
 	}
+	dialect.CompactSQL = `
+		DELETE FROM kine AS kv
+		USING	(
+			SELECT kp.prev_revision AS id
+			FROM kine AS kp
+			WHERE
+				kp.name != 'compact_rev_key' AND
+				kp.prev_revision != 0 AND
+				kp.id <= $1
+			UNION
+			SELECT kd.id AS id
+			FROM kine AS kd
+			WHERE
+				kd.deleted != 0 AND
+				kd.id <= $2
+		) AS ks
+		WHERE kv.id = ks.id`
 	dialect.TranslateErr = func(err error) error {
 		if err, ok := err.(*pq.Error); ok && err.Code == "23505" {
 			return server.ErrKeyExists