--- a/pkg/drivers/sqlite/sqlite.go
+++ b/pkg/drivers/sqlite/sqlite.go
@@ -62,22 +62,22 @@ func NewVariant(ctx context.Context, driverName, dataSourceName string, connPool
 
 	dialect.LastInsertID = true
 	dialect.CompactSQL = `
-			DELETE FROM kine AS kv
-			WHERE
-				kv.id IN (
-					SELECT kp.prev_revision AS id
-					FROM kine AS kp
-					WHERE
-						kp.name != 'compact_rev_key' AND
-						kp.prev_revision != 0 AND
-						kp.id <= ?
-					UNION
-					SELECT kd.id AS id
-					FROM kine AS kd
-					WHERE
-						kd.deleted != 0 AND
-						kd.id <= ?
-					)`
+		DELETE FROM kine AS kv
+		WHERE
+			kv.name != 'compact_rev_key' AND
+			kv.id IN (
+				SELECT kp.prev_revision AS id
+				FROM kine AS kp
+				WHERE
+					kp.prev_revision != 0 AND
+					kp.id <= ?
+				UNION
+				SELECT kd.id AS id
+				FROM kine AS kd
+				WHERE
+					kd.deleted != 0 AND
+					kd.id <= ?
+			)`
 	dialect.TranslateErr = func(err error) error {
 		if err, ok := err.(sqlite3.Error); ok && err.ExtendedCode == sqlite3.ErrConstraintUnique {
 			return server.ErrKeyExists