… go.sum
… pkg/server/compact.go