Adds additional indexes and a transaction-locked optimized native SQL compaction query that replaces the existing unlocked row-by-row compaction logic. Highly inspired by https://github.com/kubernetes/kubernetes/blob/master/staging/src/k8s.io/apiserver/pkg/storage/etcd3/compact.go

This results in a roughly **400x** reduction in compaction time for postgres, **150x** for mysql, and a **40x** for sqlite. Most of this is due to reduced round-trips and throughput to the database - the current code retrieves, evaluates, and (potentially) deletes each row individually during compaction.

Also updates SQL drivers to fix transaction handling bugs:
* github.com/lib/pq v1.1.1 => v1.8.0
* github.com/mattn/go-sqlite3 v1.10.0 => v1.14.4 (SQLite 3.29.0 => 3.33.0)

This is related to https://github.com/rancher/k3s/issues/1311

#### Test cluster: k3s 1.19.3 - 1 server 1 agent - rancher 2.5 - artifactory-jcr app
1. start kine standalone binary
1. start k3s server with external kine as datastore endpoint
1. start k3s agent
1. install cert-manager helm chart from cli
1. install rancher helm chart from cli
1. log in to rancher, configure local cluster, add artifactory-jcr app with ClusterIP
1. wait until 3rd compact is run

Code from this PR:
```
postgres: 
time="2020-10-22T16:09:44-07:00" level=info msg="Configuring DB connection pooling: maxIdleConns=2, maxOpenConns=0, connMaxLifetime=0s"
time="2020-10-22T16:09:44-07:00" level=info msg="Kine listening on tcp://0.0.0.0:2379"
time="2020-10-22T16:14:44-07:00" level=info msg="COMPACT deleted 0 rows from 2 revisions in 1.087702ms - compacted to 2/4983"
time="2020-10-22T16:19:45-07:00" level=info msg="COMPACT deleted 3084 rows from 4981 revisions in 26.46531ms - compacted to 4983/7537"
time="2020-10-22T16:24:44-07:00" level=info msg="COMPACT deleted 2427 rows from 2554 revisions in 16.755748ms - compacted to 7537/10119"

mysql:
time="2020-10-22T16:36:33-07:00" level=info msg="Configuring DB connection pooling: maxIdleConns=2, maxOpenConns=0, connMaxLifetime=0s"
time="2020-10-22T16:36:34-07:00" level=info msg="Kine listening on tcp://0.0.0.0:2379"
time="2020-10-22T16:41:34-07:00" level=info msg="COMPACT deleted 0 rows from 2 revisions in 994.448µs - compacted to 2/4832"
time="2020-10-22T16:46:34-07:00" level=info msg="COMPACT deleted 2973 rows from 4830 revisions in 77.697576ms - compacted to 4832/7550"
time="2020-10-22T16:51:34-07:00" level=info msg="COMPACT deleted 2545 rows from 2718 revisions in 42.397822ms - compacted to 7550/9896"

sqlite:
time="2020-10-22T16:53:49-07:00" level=info msg="Configuring DB connection pooling: maxIdleConns=2, maxOpenConns=0, connMaxLifetime=0s"
time="2020-10-22T16:53:49-07:00" level=info msg="Kine listening on tcp://0.0.0.0:2379"
time="2020-10-22T16:58:49-07:00" level=info msg="COMPACT deleted 0 rows from 2 revisions in 133.642µs - compacted to 2/4682"
time="2020-10-22T17:03:49-07:00" level=info msg="COMPACT deleted 2871 rows from 4680 revisions in 23.521617ms - compacted to 4682/7546"
time="2020-10-22T17:08:49-07:00" level=info msg="COMPACT deleted 2642 rows from 2864 revisions in 19.711813ms - compacted to 7546/9892"
```

Current code (with stats logging patch https://github.com/brandond/kine/commit/45327c4f0051297309ae1f8ea848e132a172eab7)
```
postgres:
time="2020-10-23T01:40:38-07:00" level=info msg="Configuring DB connection pooling: maxIdleConns=2, maxOpenConns=0, connMaxLifetime=0s"
time="2020-10-23T01:40:38-07:00" level=info msg="Kine listening on tcp://0.0.0.0:2379"
time="2020-10-23T01:45:38-07:00" level=info msg="COMPACT deleted 0 rows from 0 revisions in 321ns - compacted to 0/5071"
time="2020-10-23T01:50:48-07:00" level=info msg="COMPACT deleted 2345 rows from 4072 revisions in 9.309559208s - compacted to 4072/7584"
time="2020-10-23T01:55:46-07:00" level=info msg="COMPACT deleted 2299 rows from 2513 revisions in 7.371245154s - compacted to 6585/10132"

mysql:
time="2020-10-23T01:57:53-07:00" level=info msg="Configuring DB connection pooling: maxIdleConns=2, maxOpenConns=0, connMaxLifetime=0s"
time="2020-10-23T01:57:53-07:00" level=info msg="Kine listening on tcp://0.0.0.0:2379"
time="2020-10-23T02:02:53-07:00" level=info msg="COMPACT deleted 0 rows from 0 revisions in 491ns - compacted to 0/5269"
time="2020-10-23T02:08:03-07:00" level=info msg="COMPACT deleted 2488 rows from 4270 revisions in 10.639137845s - compacted to 4270/7635"
time="2020-10-23T02:13:00-07:00" level=info msg="COMPACT deleted 2114 rows from 2366 revisions in 7.556593544s - compacted to 6636/9972"

sqlite:
time="2020-10-23T01:20:01-07:00" level=info msg="Configuring DB connection pooling: maxIdleConns=2, maxOpenConns=0, connMaxLifetime=0s"
time="2020-10-23T01:20:01-07:00" level=info msg="Kine listening on tcp://0.0.0.0:2379"
time="2020-10-23T01:25:01-07:00" level=info msg="COMPACT deleted 0 rows from 0 revisions in 280ns - compacted to 0/5177"
time="2020-10-23T01:30:02-07:00" level=info msg="COMPACT deleted 2409 rows from 4178 revisions in 1.047498734s - compacted to 4178/7431"
time="2020-10-23T01:35:02-07:00" level=info msg="COMPACT deleted 1994 rows from 2254 revisions in 697.222053ms - compacted to 6432/9780"
```