@davidnuzik 2020-10-22
When this is ready, can we also backport into 1.18 and 1.19 ? Hosted Rancher will be needing this fix.