@brandond 2020-10-23
Here are some perf figures for the pure-sql compact. I'm working on getting some perf figures for the current mechanism and will update later this evening.

```
k3s 1.19.3 - 1 server 1 agent - rancher 2.5 - artifactory-jcr app

postgres: 
time="2020-10-22T16:09:44-07:00" level=info msg="Configuring DB connection pooling: maxIdleConns=2, maxOpenConns=0, connMaxLifetime=0s"
time="2020-10-22T16:09:44-07:00" level=info msg="Kine listening on tcp://0.0.0.0:2379"
time="2020-10-22T16:14:44-07:00" level=info msg="COMPACT deleted 0 rows from 2 revisions in 1.087702ms - compacted to 2/4983"
time="2020-10-22T16:19:45-07:00" level=info msg="COMPACT deleted 3084 rows from 4981 revisions in 26.46531ms - compacted to 4983/7537"
time="2020-10-22T16:24:44-07:00" level=info msg="COMPACT deleted 2427 rows from 2554 revisions in 16.755748ms - compacted to 7537/10119"

mysql:
time="2020-10-22T16:36:33-07:00" level=info msg="Configuring DB connection pooling: maxIdleConns=2, maxOpenConns=0, connMaxLifetime=0s"
time="2020-10-22T16:36:34-07:00" level=info msg="Kine listening on tcp://0.0.0.0:2379"
time="2020-10-22T16:41:34-07:00" level=info msg="COMPACT deleted 0 rows from 2 revisions in 994.448µs - compacted to 2/4832"
time="2020-10-22T16:46:34-07:00" level=info msg="COMPACT deleted 2973 rows from 4830 revisions in 77.697576ms - compacted to 4832/7550"
time="2020-10-22T16:51:34-07:00" level=info msg="COMPACT deleted 2545 rows from 2718 revisions in 42.397822ms - compacted to 7550/9896"

sqlite:
time="2020-10-22T16:53:49-07:00" level=info msg="Configuring DB connection pooling: maxIdleConns=2, maxOpenConns=0, connMaxLifetime=0s"
time="2020-10-22T16:53:49-07:00" level=info msg="Kine listening on tcp://0.0.0.0:2379"
time="2020-10-22T16:58:49-07:00" level=info msg="COMPACT deleted 0 rows from 2 revisions in 133.642µs - compacted to 2/4682"
time="2020-10-22T17:03:49-07:00" level=info msg="COMPACT deleted 2871 rows from 4680 revisions in 23.521617ms - compacted to 4682/7546"
time="2020-10-22T17:08:49-07:00" level=info msg="COMPACT deleted 2642 rows from 2864 revisions in 19.711813ms - compacted to 7546/9892"
```